+++ org-roam.el
@@ -58,6 +58,13 @@ All Org files, at any level of nesting, is considered part of the Org-roam."
   :type 'directory
   :group 'org-roam)
 
+(defcustom org-roam-new-file-directory nil
+  "Path to where new Org-roam files are created.
+
+If nil, default to the org-roam-directory (preferred)."
+  :type 'directory
+  :group 'org-roam)
+
 (defcustom org-roam-buffer-position 'right
   "Position of `org-roam' buffer.
 
@@ -172,11 +179,13 @@ If called interactively, then PARENTS is non-nil."
     (org-roam--build-cache-async)
     (user-error "Your Org-Roam cache isn't built yet! Please wait")))
 
-(defun org-roam--org-roam-file-p ()
-  "Return t if file is part of org-roam system, false otherwise."
-  (and (buffer-file-name (current-buffer))
-       (f-descendant-of-p (file-truename (buffer-file-name (current-buffer)))
-                          (file-truename org-roam-directory))))
+(defun org-roam--org-roam-file-p (&optional file)
+  "Return t if FILE is part of org-roam system, defaulting to the name of the current buffer. Else, return nil."
+  (let ((path (or file
+                  (buffer-file-name (current-buffer)))))
+    (and path
+         (f-descendant-of-p (file-truename path)
+                            (file-truename org-roam-directory)))))
 
 (defun org-roam--get-title-from-cache (file)
   "Return title of `FILE' from the cache."
@@ -199,7 +208,8 @@ If `ABSOLUTE', return an absolute file-path. Else, return a relative file-path."
                               (if org-roam-encrypt-files
                                   (concat id ".org.gpg")
                                 (concat id ".org"))
-                              org-roam-directory))))
+                              (or org-roam-new-file-directory
+                                  org-roam-directory)))))
     (if absolute
         absolute-file-path
       (file-relative-name absolute-file-path
@@ -250,9 +260,9 @@ If not provided, derive the title from the file name."
   "Create an org-roam file at FILE-PATH, optionally setting the TITLE attribute."
   (if (file-exists-p file-path)
       (error (format "Aborting, file already exists at %s" file-path))
+    (make-empty-file file-path t)
     (if org-roam-autopopulate-title
-        (org-roam--populate-title file-path title)
-      (make-empty-file file-path))
+        (org-roam--populate-title file-path title))
     (save-excursion
       (with-current-buffer (find-file-noselect file-path)
         (org-roam--update-cache)))))
@@ -296,7 +306,8 @@ If PREFIX, downcase the title before insertion."
          (title (completing-read "File: " completions))
          (absolute-file-path (or (cadr (assoc title completions))
                                  (org-roam--make-new-file-path (org-roam--get-new-id title) t)))
-         (current-file-path (-> (current-buffer)
+         (current-file-path (-> (or (buffer-base-buffer)
+                                    (current-buffer))
                                 (buffer-file-name)
                                 (file-truename)
                                 (file-name-directory))))
@@ -324,6 +335,26 @@ If PREFIX, downcase the title before insertion."
       (org-roam--make-file absolute-file-path title-or-slug))
     (find-file absolute-file-path)))
 
+(defun org-roam-switch-to-buffer ()
+  "Switch to an existing org-roam buffer using completing-read."
+  (interactive)
+  (let* ((all-buffers (buffer-list))
+         (roam-buffers
+          (--filter (and (with-current-buffer it (derived-mode-p 'org-mode))
+                         (buffer-file-name it)
+                         (org-roam--org-roam-file-p (buffer-file-name it)))
+                    all-buffers))
+         (names-and-buffers (mapcar (lambda (buffer)
+                                      (cons (or (org-roam--get-title-from-cache
+                                                 (buffer-file-name buffer))
+                                                (buffer-name buffer))
+                                            buffer))
+                                    roam-buffers)))
+    (unless roam-buffers
+      (error "No roam buffers."))
+    (when-let ((name (completing-read "Choose a buffer: " names-and-buffers)))
+      (switch-to-buffer (cdr (assoc name names-and-buffers))))))
+
 ;;; Building the org-roam cache
 (defun org-roam--build-cache-async ()
   "Builds the caches asychronously."
@@ -350,11 +381,13 @@ If PREFIX, downcase the title before insertion."
   (setq org-roam-backward-links-cache (make-hash-table :test #'equal))
   (setq org-roam-titles-cache (make-hash-table :test #'equal)))
 
-(defun org-roam--clear-file-from-cache ()
+(defun org-roam--clear-file-from-cache (&optional filepath)
   "Remove any related links to the file.
 
 This is equivalent to removing the node from the graph."
-  (let ((file (file-truename (buffer-file-name (current-buffer)))))
+  (let* ((path (or filepath
+                   (buffer-file-name (current-buffer))))
+         (file (file-truename path)))
     ;; Step 1: Remove all existing links for file
     (when-let ((forward-links (gethash file org-roam-forward-links-cache)))
       ;; Delete backlinks to file
@@ -604,6 +637,50 @@ This needs to be quick/infrequent, because this is run at
       (insert graph))
     (call-process org-roam-graphviz-executable nil 0 nil temp-dot "-Tsvg" "-o" temp-graph)
     (call-process org-roam-graph-viewer nil 0 nil temp-graph)))
+
+(defun org-roam--rename-file-links (file new-file &rest args)
+  "Rename backlinks of FILE to refer to NEW-FILE."
+  (when (and (not (auto-save-file-name-p file))
+             (not (auto-save-file-name-p new-file))
+             (org-roam--org-roam-file-p new-file))
+    (org-roam--ensure-cache-built)
+    (org-roam--clear-file-from-cache file)
+
+    (let* ((files (gethash file org-roam-backward-links-cache nil))
+           (path (file-truename file))
+           (new-path (file-truename new-file))
+           (slug (org-roam--get-title-or-slug file))
+           (old-title (format org-roam-link-title-format slug))
+           (new-slug (or (org-roam--get-title-from-cache path)
+                         (org-roam--get-title-or-slug new-path)))
+           (new-title (format org-roam-link-title-format new-slug)))
+      (when files
+        (maphash (lambda (file-from props)
+                   (let* ((file-dir (file-name-directory file-from))
+                          (relative-path (file-relative-name new-path file-dir))
+                          (old-relative-path (file-relative-name path file-dir))
+                          (slug-regex (regexp-quote (format "[[file:%s][%s]]" old-relative-path old-title)))
+                          (named-regex (concat
+                                        (regexp-quote (format "[[file:%s][" old-relative-path))
+                                        "\\(.*\\)"
+                                        (regexp-quote "]]"))))
+                     (with-temp-file file-from
+                       (insert-file-contents file-from)
+                       (while (re-search-forward slug-regex  nil t)
+                         (replace-match (format "[[file:%s][%s]]" relative-path new-title)))
+                       (goto-char (point-min))
+                       (while (re-search-forward named-regex nil t)
+                         (replace-match (format "[[file:%s][\\1]]" relative-path))))
+                     (save-window-excursion
+                       (find-file file-from)
+                       (org-roam--update-cache))))
+                 files))
+      (save-window-excursion
+        (find-file new-path)
+        (org-roam--update-cache)))))
+
+(advice-add 'rename-file :after 'org-roam--rename-file-links)
+(advice-add 'delete-file :before 'org-roam--clear-file-from-cache)
 
 (provide 'org-roam)